When finding the end of paragraphs, account for backticks within HTML attributes
For example:
    <span attr="has ` in quotes"><pre><code attr="here ` too">

The previous algorithm would filter off the "<pre>" thinking that
it's within a code-span, but it actually isn't.

This commit handles this case correctly by switching priority between
the code-span detection algo and the HTML parsing algo depending on the
context.

Even with this commit, the algo does not handle backticks within
HTML attribute names, but it should be ok to not support that.

Fixes vfmd/vfmd-spec#1