Automatic links: Fix regexp for detecting email addresses as per RFC 5322
According to RFC 5322, an email address is of the form `local-part@domain`,
where both local-part and domain should not contain: `()<>[]:;@\,."` and
ASCII whitespace / control characters.

We add `` ` `` to the list so that code-spans get recognized with a
higher priority.

We also don't worry about the obscure cases (quoted-strings and obsolete
addressing).