Identifying reference resolution blocks should handle images within the id
For example, if I want an image inside a link, I would write:

    Here's a [![picture]] you can click to go to my site.

    [picture]: path/to/picture.png
    [![picture]]: http://mysite.com

Previously, unescaped `[` and `]` inside a ref meant that the second
ref-resolution block in the above example was not even identified. This
commit changes that to allow for images to be present in the ref-id of
a ref-resolution block.